Reapply "[lit] Implement ulimit builtin"
This reverts commit e81214230d3621bfeab02096e932c4417783aa17.

This was causing some test failures on MacOS that are now fixed in the reland.
These failures were related to calling ulimit -v despite XNU not having support
for that option. This patch simply disables the test on non-Linux platforms for
now until we can have a Linux specific test for ulimit -v.